requirement.txt
fastapi
uvicorn
tensorflow
pillow
numpy